pyproject.toml
[project]
name = "hud-ui-cube-local"
version = "0.1.0"
requires-python = ">=3.11,<3.13"
dependencies = [
  "hud-python>=0.5.0",
  "pyautogui",
  "playwright",
  "httpx",
  "typer",
  "google-api-python-client",
  "google-auth",
  "pillow",
  "anthropic",
  "google-genai",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["tools", "setup", "scenarios"]
